requirements.txt
flask
flask-sqlalchemy
cryptography
pyotp
qrcode[pil]
authlib
msal
python-dotenv
requests